Updated to work with nim 1.6.0
Nim nolonger needs an external library for unsigned integers, so I
removed the call for that library.
I forced the gnu99 dialect of c for compatibility, I'll work on getting
a newer version to work later.
I changed i586-elf-* to i686-elf-* since that is the cross compiler I
have and what osdev says to use. I also explicitly set nimcache since
the default is outside the repo and it was less work to update the
linker command. Nim has a different naming scheme for cached object
files so I updated those in the linker command.